+++ .idea/dataSources.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="DataSourceManagerImpl" format="xml" multifile-model="true">
-    <data-source source="LOCAL" name="coffeemanager@localhost" uuid="803f656c-b400-4bb7-b186-37108cc48db0">
+    <data-source source="LOCAL" name="coffeemanager@localhost" uuid="94ce1072-2b46-4e51-b71d-b314b3d0ca0b">
       <driver-ref>mysql.8</driver-ref>
       <synchronize>true</synchronize>
       <jdbc-driver>com.mysql.cj.jdbc.Driver</jdbc-driver>
