+++ src/main/java/coffeemanager/infra/util/FileUtil.java
@@ -3,8 +3,6 @@ package coffeemanager.infra.util;
 import java.io.File;
 import java.io.IOException;
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
@@ -16,23 +14,25 @@ public class FileUtil {
     @Value("${upload.path}")
     private String filePath;
 
-    public List<FileDto> upload(List<MultipartFile> files, String depth) throws IOException {
-        List<FileDto> fileDtos = new ArrayList<>();
-
-        if (files.isEmpty() || files.getFirst().isEmpty()) {
-            return fileDtos;
+    // 저장하기전
+    public FileDto upload(MultipartFile file, String depth) throws IOException {
+        //파일 비어있는지 확인
+        if(file.isEmpty()){
+            throw new IllegalArgumentException("File is empty");
         }
+
+        //파일명, 저장 경로
+        String originFileName = file.getOriginalFilename();
+        String renameFileName = generateRenameFileName(originFileName);
         String savePath = createSavePath(depth);
 
-        for (MultipartFile file : files) {
-            String originFileName = file.getOriginalFilename();
-            String renameFileName = generateRenameFileName(originFileName);
-            FileDto fileDto = new FileDto(originFileName, renameFileName, savePath);
-            fileDtos.add(fileDto);
-            uploadFile(file, fileDto);
-        }
+        FileDto fileDto = new FileDto(originFileName, renameFileName, savePath);
 
-        return fileDtos;
+        //실제 저장 아래 uploadFile 로
+        uploadFile(file, fileDto);
+
+        // 저장후 fileDto 반환
+        return fileDto;
     }
 
     private void uploadFile(MultipartFile file, FileDto fileDto) throws IOException {
@@ -41,7 +41,7 @@ public class FileUtil {
             path.mkdirs();
         }
 
-        File target = new File(filePath + fileDto.savePath() + fileDto.renameFilename());
+        File target = new File(filePath + fileDto.savePath() + fileDto.renameFileName());
         file.transferTo(target);
     }
 
@@ -52,7 +52,8 @@ public class FileUtil {
 
     private String createSavePath(String depth) {
         LocalDate now = LocalDate.now();
-        return depth + "/" +
+        String pathDepth = (depth == null || depth.isEmpty()) ? "" : depth + "/";
+        return pathDepth + "/" +
             now.getYear() + "/" +
             now.getMonth() + "/" +
             now.getDayOfMonth() + "/";
