--- src/main/java/coffeemanager/util/file/FileDto.java
@@ -1,9 +1,0 @@
-package coffeemanager.util.file;
-
-public record FileDto(
-    String originFileName,
-    String renameFileName,
-    String savePath
-) {
-
-}
