+++ src/main/resources/application.properties
@@ -5,5 +5,5 @@ spring.datasource.password=123qwe!@#
 spring.datasource.driver=com.mysql.cj.jdbc.Driver
 spring.datasource.hikari.maximum-pool-size=10
 remember-me.key=123qwe!@#
-#schema.script.name=schema.sql
-#data.script.name=data.sql
+schema.script.name=schema.sql
+data.script.name=data.sql
